More cosmetic changes to point in polygon code.

+++ test/Itinero.Test/LocalGeo/PointInPolygonTest.cs
@@ -38,10 +38,10 @@ namespace Itinero.Test.LocalGeo
         [Test]
         public void TestPointInPolygon()
         {
-            var c1 = new Coordinate((float)51.2157978, (float)3.2200444);
-            var c2 = new Coordinate((float)51.2157465, (float)3.2200733);
-            var c3 = new Coordinate((float)51.2157229, (float)3.2199664);
-            var c4 = new Coordinate((float)51.2157743, (float)3.2199375);
+            var c1 = new Coordinate(51.2157978f, 3.2200444f);
+            var c2 = new Coordinate(51.2157465f, 3.2200733f);
+            var c3 = new Coordinate(51.2157229f, 3.2199664f);
+            var c4 = new Coordinate(51.2157743f, 3.2199375f);
 
             var p = new Polygon();
             p.ExteriorRing.Add(c1);
@@ -50,14 +50,14 @@ namespace Itinero.Test.LocalGeo
             p.ExteriorRing.Add(c4);
 
             // In the middle of the polygon
-            var testPoint0 = new Coordinate((float)51.2157635, (float)3.2200099);
+            var testPoint0 = new Coordinate(51.2157635f, 3.2200099f);
             Assert.IsTrue(p.PointIn(testPoint0));
             // just outside of the polygon, but within the bounding box
-            var testPoint1 = new Coordinate((float)51.2157898, (float)3.2200648);
+            var testPoint1 = new Coordinate(51.2157898f, 3.2200648f);
             Assert.IsFalse(p.PointIn(testPoint1));
 
             // way outside of the poly
-            var testPoint2 = new Coordinate((float)51.2158183, (float)3.2201524);
+            var testPoint2 = new Coordinate(51.2158183f, 3.2201524f);
             Assert.IsFalse(p.PointIn(testPoint2));
         }
     }
